[SPARK-45546][BUILD][INFRA] Make `publish-snapshot` support `package` first then `deploy`
### What changes were proposed in this pull request?
This pr adds an environment variable `PACKAGE_BEFORE_DEPLOY` to the `publish-snapshot` process. When `PACKAGE_BEFORE_DEPLOY` is true, the publish process will be split into two steps: the first step is to package with mvn package, and the second step is to deploy the packaged jar.

At the same time, this PR sets `PACKAGE_BEFORE_DEPLOY` to true in the `publish_snapshot.yml` configuration.

### Why are the changes needed?
Make the `publish-snapshot` task in GitHub Action to be divided into two steps, which can alleviate the resource pressure brought by direct deploy.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass GitHub Actions

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43378 from LuciferYang/no-doc-deploy.

Lead-authored-by: yangjie01 <yangjie01@baidu.com>
Co-authored-by: YangJie <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>