[SPARK-45009][SQL][FOLLOW UP] Turn off decorrelation in join conditions for AQE InSubquery test
### What changes were proposed in this pull request?

Turn off decorrelation for predicate subquery in join conditions for an insubquery test in AdaptiveQueryExecSuite.

### Why are the changes needed?

The test depends on an assumption that the query won't be decorrelated to test that InSubquery is allowed during AQE. Disabling the decorrelation for predicate subqueries in join condition feature so it doesn't silently lose test coverage.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Test only change.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43408 from andylam-db/insubquery-test-fix.

Authored-by: Andy Lam <andy.lam@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>