[SPARK-45541][CORE] Add SSLFactory
### What changes were proposed in this pull request?

As titled - add a factory which supports creating SSL engines, and a corresponding builder for it. This will be used in a follow up PR by the `TransportContext` and related files to add SSL support.

### Why are the changes needed?

We need a mechanism to initialize the appropriate SSL implementation with the configured settings (such as protocol, ciphers, etc) for RPC SSL support.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

existing tests. This will be more thoroughly tested in a follow up PR which adds callsites to it. It has been integration tested as part of https://github.com/apache/spark/pull/42685

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43386 from hasnain-db/spark-tls-factory.

Authored-by: Hasnain Lakhani <hasnain.lakhani@databricks.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>