[SPARK-45209][CORE][UI] Flame Graph Support For Executor Thread Dump Page
### What changes were proposed in this pull request?

This PR draws a CPU Flame Graph by Java stack traces for executors and drivers. Currently, the Java stack traces is just a SNAPSHOT, not sampling at a certain frequency for a period. Sampling might be considered an upcoming feature out of the scope of this PR.

![fg git](https://github.com/apache/spark/assets/8326978/c3f99a1a-78ee-4adb-be1f-e4afd5f307b7)

If you are new to flame graphs, there are also some references you can refer to learn about the basic concepts and details.

[1] [Flame Graphs](https://www.brendangregg.com/flamegraphs.html)
[2] [FLIP-165: Operator's Flame Graphs](https://cwiki.apache.org/confluence/display/FLINK/FLIP-165%3A+Operator%27s+Flame+Graphs)
[3] [Java in Flames. mixed-mode flame graphs provide a… | by Netflix Technology Blog](https://netflixtechblog.com/java-in-flames-e763b3d32166)
[4] [HProf](https://docs.oracle.com/javase/7/docs/technotes/samples/hprof.html)

#### Pending features

This PR mainly focuses on the UI, independent of the profiling steps. What we might have in the future are:

- Flame Graph Support For Task Thread Page which SPARK-45151 added
- Add `ProfilingExecutor(max, interval)` message to profile whole executor
- Add `ProfileTask(taskId, max, interval)` message to profile an certain task
- Different views for on/off/full CPUs
- Mixed mode profiling, which might rely upon some ext libs at runtime
- And so on.

### Why are the changes needed?

Performance is always an important design factor in Spark. It is desirable to provide better visibility into the distribution of CPU resources while executing user code alongside the Spark kernel. One of the most visually effective means to do that is [Flame Graphs](http://www.brendangregg.com/FlameGraphs/cpuflamegraphs.html), which visually presents the data gathered by performance profiling tools used by developers for performance tuning their applications.

### Does this PR introduce _any_ user-facing change?

yes

### How was this patch tested?

locally
### Was this patch authored or co-authored using generative AI tooling?

no

Closes #42988 from yaooqinn/SPARK-45209.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: yangjie01 <yangjie01@baidu.com>