[SPARK-45651][BUILD][FOLLOWUP] Reduce mvn -Xmx option to 2g in publish_snapshot workflow
### What changes were proposed in this pull request?
Limit max memory for `mvn clean deploy` to `2g` when run in `publish_snapshot` Github workflow.

### Why are the changes needed?
The host that runs the workflow has only 7G of memory, while the `release-build.sh` script sets the limit to 12g, causing the process to be killed (for branch `master`).

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Not tested

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43538 from EnricoMi/publish-snapshot-mvn-xmx.

Authored-by: Enrico Minack <github@enrico.minack.dev>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>