[SPARK-45791][CONNECT][TESTS] Rename `SparkConnectSessionHodlerSuite.scala` to `SparkConnectSessionHolderSuite.scala`
### What changes were proposed in this pull request?

This PR aims to fix a typo `Hodler` in file name.
- `SparkConnectSessionHodlerSuite.scala` (from)
- `SparkConnectSessionHolderSuite.scala` (to)

It's also unmatched with the class name in the file because class name itself is correct.

https://github.com/apache/spark/blob/ca9d508f6005554f389173f591721b7a949e45b1/connector/connect/server/src/test/scala/org/apache/spark/sql/connect/service/SparkConnectSessionHodlerSuite.scala#L37

### Why are the changes needed?

This is a typo from the original PR.
- https://github.com/apache/spark/pull/41580

Since the original PR is shipped as Apache Spark 3.5.0, I created a JIRA instead of a follow-up. We need to backport this patch to `branch-3.5`.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43657 from dongjoon-hyun/SPARK-45791.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>