[SPARK-45808][CONNECT][PYTHON] Better error handling for SQL Exceptions
### What changes were proposed in this pull request?
This patch optimizes the handling of errors reported back to Python. First, it properly allows the extraction of the `ERROR_CLASS` and the `SQL_STATE` and gives simpler accces to the stack trace.

It therefore makes sure that the display of the stack trace is no longer only server-side decided but becomes a local usability property.

In addition the following methods on the `SparkConnectGrpcException` become actually useful:

* `getSqlState()`
* `getErrorClass()`
* `getStackTrace()`

### Why are the changes needed?
Compatibility

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Updated the existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43667 from grundprinzip/SPARK-XXXX-ex.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>