[SPARK-45431][DOCS] Document new SSL RPC feature
### What changes were proposed in this pull request?

Add some documentation to clarify the new flags being added and how this feature interacts with existing SSL and encryption support. It's a little confusing so feedback is welcome.

Note I am not sure if it's best practice to merge this after the feature is fully merged - thought I would put this up now since it's unblocked.

### Why are the changes needed?

New features require documentation so users can understand how to use them.

### Does this PR introduce _any_ user-facing change?

Yes, this is adding documentation for a new feature

### How was this patch tested?

Not applicable, this is documentation

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43240 from hasnain-db/spark-tls-docs.

Authored-by: Hasnain Lakhani <hasnain.lakhani@databricks.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>