[SPARK-45794][SS] Introduce state metadata source to query the streaming state metadata information
### What changes were proposed in this pull request?

Introduce a new data source so that user can query the metadata of each state store of a streaming query, the schema of the result will be following:

column | type
--------- | -------
| operatorId | INT |
| operatorName | STRING |
| stateStoreName | STRING |
| numPartitions | INT |
| minBatchId | LONG |
| minBatchId | LONG |
| _numColsPrefixKey (metadata column) | INT |

To use this source, specify the source format and checkpoint path and load the dataframe

`df = spark.read.format(“state-metadata”).load(“/checkpointPath”)`

### Why are the changes needed?

To improve debugability. Also facilitate the query of state store data source introduced in SPARK-45511 by displaying the operator id, batch id and state store name.

### Does this PR introduce _any_ user-facing change?

Yes, this is a new source exposed to user.

### How was this patch tested?

Add test to verify the output of state metadata

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43660 from chaoqin-li1123/state_metadata_source.

Authored-by: Chaoqin Li <chaoqin.li@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>