[SPARK-45752][SQL] Unreferenced CTE should all be checked by CheckAnalysis0
### What changes were proposed in this pull request?

This PR fixes an issue that if a CTE is referenced by a non-referenced CTE,  then this CTE should also have ref count as 0 and goes through CheckAnalysis0. This will guarantee analyzer throw proper error message for problematic CTE which is not referenced.

### Why are the changes needed?

To improve error message for non-referenced CTE case.

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

UT

### Was this patch authored or co-authored using generative AI tooling?

NO

Closes #43614 from amaliujia/cte_ref.

Lead-authored-by: Rui Wang <rui.wang@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>