[SPARK-45915][SQL] Treat decimal(x, 0) the same as IntegralType in `PromoteStrings`
### What changes were proposed in this pull request?

The common type of decimal(x, 0) and string is double. But the common type of int/bigint and string are int/bigint.

This PR updates `PromoteStrings` make the common type of decimal(x, 0) and string is decimal(x, 0).

### Why are the changes needed?

1. Make decimal(x, 0) behave the same as int/bigint in `PromoteStrings`.
2. Reduce one cast in binary comparison so we may use bucket read. For example: `cast(stringCol as double) = cast(decimalCol as double)` vs `cast(stringCol as decimal(x, 0)) = decimalCol`.

### Does this PR introduce _any_ user-facing change?

Yes. The result type of decimal(x, 0) and string is decimal(x, 0) in binary comparison.

### How was this patch tested?

Unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43812 from wangyum/SPARK-45915.

Authored-by: Yuming Wang <yumwang@ebay.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>