[SPARK-45810][PYTHON] Create Python UDTF API to stop consuming rows from the input table
### What changes were proposed in this pull request?

This PR creates a Python UDTF API to stop consuming rows from the input table.

If the UDTF raises a `SkipRestOfInputTableException` exception in the `eval` method, then the UDTF stops consuming rows from the input table for that input partition, and finally calls the `terminate` method (if any) to represent a successful UDTF call.

For example:

```
udtf(returnType="total: int")
class TestUDTF:
    def __init__(self):
        self._total = 0

    def eval(self, _: Row):
        self._total += 1
        if self._total >= 3:
            raise SkipRestOfInputTableException("Stop at self._total >= 3")

    def terminate(self):
        yield self._total,
```

### Why are the changes needed?

This is useful when the UDTF logic knows that we don't have to scan the input table anymore, and skip the rest of the I/O for that case.

### Does this PR introduce _any_ user-facing change?

Yes, see above.

### How was this patch tested?

This PR adds test coverage.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43682 from dtenedor/udtf-api-stop-consuming-input-rows.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Takuya UESHIN <ueshin@databricks.com>