[SPARK-45827][SQL] Fix variant parquet reader
## What changes were proposed in this pull request?

This is a follow-up of https://github.com/apache/spark/pull/43707. The previous PR missed a piece in the variant parquet reader: we are treating the variant type as `struct<value binary, metadata binary>`, so it also needs a similar `assembleStruct` process in the Parquet reader to correctly set the nullness of variant values from def/rep levels.

## How was this patch tested?

Extend the existing unit test. It would fail without the change.

Closes #43825 from chenhao-db/fix_variant_parquet_reader.

Authored-by: Chenhao Li <chenhao.li@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>