[SPARK-45747][SS] Use prefix key information in state metadata to handle reading state for session window aggregation
### What changes were proposed in this pull request?
Currently reading state for session window aggregation operator is not supported because the numColPrefixKey is unknown. We can read the operator state metadata introduced in SPARK-45558 to determine the number of prefix columns and load the state of session window correctly.

### Why are the changes needed?
To support reading state for session window aggregation.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Add integration test.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #43788 from chaoqin-li1123/session_window_agg.

Authored-by: Chaoqin Li <chaoqin.li@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>