[SPARK-45938][INFRA] Add `utils` to the dependencies of the `core/unsafe/network_common` module in `module.py`
### What changes were proposed in this pull request?
This pr add `utils` to the dependencies of the `core/unsafe/network_common` module in `module.py` due to  `utils` is direct dependency of `core/unsafe/network_common`:

https://github.com/apache/spark/blob/60b507b90e4a083f4452777cb17f52d81eba18ae/core/pom.xml#L102-L106

https://github.com/apache/spark/blob/33b337ab75369f41bce87f3dc5cba69fabf2b259/common/unsafe/pom.xml#L44-L48

https://github.com/apache/spark/blob/33b337ab75369f41bce87f3dc5cba69fabf2b259/common/network-common/pom.xml#L177-L181

### Why are the changes needed?
All UTs of `utils` module are still in the `core` module now. This pr is used to avoid the issue of not running unit tests in the `core` module when changing the `utils` module.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass GitHub Actions

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43818 from LuciferYang/common-utils-test.

Lead-authored-by: yangjie01 <yangjie01@baidu.com>
Co-authored-by: YangJie <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>