[SPARK-45960][INFRA] Add Python 3.10 to the Daily Python Github Action job
### What changes were proposed in this pull request?

This PR aims to enable `Python 3.10` testing in the following daily `Python-only` Github Action job.

https://github.com/apache/spark/actions/workflows/build_python.yml

### Why are the changes needed?

To provide `Python 3.10` test coverage to Apache Spark 4.0.0.

Since SPARK-45953 installed `Python 3.10` into the infra image, what we need is to add it to the daily job.
- #43840

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

We need to validate this in the daily GitHub Action job.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43847 from dongjoon-hyun/SPARK-45960.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>