[SPARK-45827][SQL] Use StructNullableTypeConverter for Variant
### What changes were proposed in this pull request?

Add a small fix for #43707. Since Variant is represented in columnar form as a struct, it must use `StructNullableTypeConverter` so that nulls are set properly in child column vectors.

### Why are the changes needed?

Fixes a potential when setting nulls in Variant columns.

### Does this PR introduce _any_ user-facing change?

No, Variant is not released yet.

### How was this patch tested?

Updated existing unit test to test Variant. It fails without the fix.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43911 from cashmand/SPARK-45827-fixnulls.

Authored-by: cashmand <david.cashman@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>