[SPARK-46099][PS][DOCS] Refactor "Supported pandas API" generation script
### What changes were proposed in this pull request?

This PR proposes to refactor the script used to generate the [Supported pandas API](https://spark.apache.org/docs/latest/api/python/user_guide/pandas_on_spark/supported_pandas_api.html) documentation. The script has been restructured for better readability and maintainability. The refactoring includes:

- Simplifying complex functions and breaking them into smaller, more manageable pieces.
- Improving variable and function naming for clarity.
- Adding comprehensive docstrings in the NumPy docstyle.
- Streamlining the flow of the script to enhance logical coherence.

### Why are the changes needed?

The previous version of the script was hard to understand and maintain due to its complexity and lack of documentation. This refactoring makes the script more accessible to new contributors and easier to modify or extend in the future. It also ensures that the script adheres to best practices in Python coding, making it a more reliable tool for generating accurate and up-to-date documentation.

### Does this PR introduce _any_ user-facing change?

No user-facing changes. This PR only affects the internal documentation generation process.

### How was this patch tested?

Tested by generating the documentation manually and verifying that the output remains consistent with the previous version.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44010 from itholic/refactor_doc_gen_script.

Authored-by: Haejoon Lee <haejoon.lee@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>