[SPARK-46111][DOCS][PYTHON] Add copyright to the PySpark official documentation
### What changes were proposed in this pull request?

This PR proposes to add the Apache Spark Foundation copyright notice to the bottom of the PySpark official documentation.

### Why are the changes needed?

Our current documentation is missing the copyright. The addition of the copyright notice is necessary to ensure compliance with the Apache Software Foundation's requirements for project documentation.

### Does this PR introduce _any_ user-facing change?

No API changes, but users will now see the Apache Spark Foundation copyright notice at the bottom of each page of the PySpark documentation as below:

## Before
<img width="278" alt="Screenshot 2023-11-27 at 11 49 38 AM" src="https://github.com/apache/spark/assets/44108233/fda0c23d-5a8c-457c-adde-08917282320f">

## After

<img width="770" alt="Screenshot 2023-11-27 at 11 35 35 AM" src="https://github.com/apache/spark/assets/44108233/bfecf5ca-7117-4269-8e85-59646ef60b70">

### How was this patch tested?

Manually build the docs and confirm.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44026 from itholic/add_copyright.

Authored-by: Haejoon Lee <haejoon.lee@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>