[SPARK-45833][SS][DOCS] Document the new introduction of state data source
### What changes were proposed in this pull request?

This PR proposes to add a new doc page describing the new data source, `state data source`.

Worth noting that we explicitly mention the data source as experimental, so that we do not close the opportunity to improve further if it's backward incompatible.

### Why are the changes needed?

The data source is an user-facing one and would be evolved over time, hence we'll need to document and publicize it.

### Does this PR introduce _any_ user-facing change?

Yes, doc change.

### How was this patch tested?

Built the docs directory with jekyll. Here are relevant screenshots.

<img width="1136" alt="스크린샷 2023-11-28 오후 5 02 47" src="https://github.com/apache/spark/assets/1317309/b36fea49-89f0-484d-b90d-be069c589b8f">

![structured-streaming-state-data-source-1](https://github.com/apache/spark/assets/1317309/c65cef9f-750b-4c00-a289-c705386a538b)

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43920 from HeartSaVioR/SPARK-45833.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>