[SPARK-46191][CORE] Improve `FileSystemPersistenceEngine.persist` error msg in case of the existing file
### What changes were proposed in this pull request?

This PR aims to improve `FileSystemPersistenceEngine.persist`'s error message in case of the existing file.

### Why are the changes needed?

`FileSystemPersistenceEngine.persist` fails when a file exists in the location where `key` points.
It's helpful for us to distinguish this error from the normal file creation failures due to the other disk issues like, READ-ONLY mount.

### Does this PR introduce _any_ user-facing change?

No. This changes only a error message for a better UX.

### How was this patch tested?

Pass the CIs with the newly added test case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44100 from dongjoon-hyun/SPARK-46191.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>